Refactor upgrade handling by modularizing upgrade functions into dedicated files and removing the obsolete upgradeHandlers module
